add integration tests for empty folder
